refactored code and restructed files
